+++ src/AABB.hpp
@@ -12,8 +12,8 @@ namespace aabb {
 
 class AABB {
 public:
-  AABB(int x1, int y1, int x2, int y2);
-  int x1, y1, x2, y2;
+  AABB(float x1, float y1, float x2, float y2);
+  float x1, y1, x2, y2;
   AABB unite(const AABB &) const;
   bool contains(const AABB &) const;
   bool overlaps(const AABB &) const;
